+++ projects/cream/index.js
@@ -2,8 +2,7 @@ const sdk = require("@defillama/sdk");
 const utils = require("../helper/utils");
 const { unwrapUniswapLPs } = require("../helper/unwrapLPs");
 const { getCompoundV2Tvl, compoundExports } = require("../helper/compound");
-const { transformFantomAddress, transformBscAddress } = require('../helper/portedTokens')
-const { GraphQLClient, gql } = require('graphql-request')
+const {  transformBscAddress } = require('../helper/portedTokens')
 
 const abiCerc20 = require("./cerc20.json");
 const abiCereth2 = require("./creth2.json");
@@ -75,42 +74,6 @@ async function ethereumTvl(timestamp, block) {
     }
   });
   await unwrapUniswapLPs(balances, lpPositions, block);
-
-  let iron_bank_tokens = (
-    await utils.fetchURL(
-      "https://api.cream.finance/api/v1/crtoken?comptroller=ironbank"
-    )
-  ).data;
-  // --- Grab all the getCash values of cyERC20 (Iron Bank)---
-  const ironBankCalls = iron_bank_tokens.map((token) => ({ target: token.token_address }))
-  let [cashValuesIronBank, totalBorrowsIB, totalReservesIB, underlyingsIronBank] = await Promise.all([
-    sdk.api.abi.multiCall({
-      block,
-      calls: ironBankCalls,
-      abi: abiCerc20["getCash"],
-    }),
-    sdk.api.abi.multiCall({
-      block,
-      calls: ironBankCalls,
-      abi: abiCerc20["totalBorrows"],
-    }),
-    sdk.api.abi.multiCall({
-      block,
-      calls: ironBankCalls,
-      abi: abiCerc20["totalReserves"],
-    }),
-    sdk.api.abi.multiCall({
-      block,
-      calls: ironBankCalls,
-      abi: abiCerc20["underlying"],
-    })
-  ]);
-
-  cashValuesIronBank.output.map((cashVal, idx) => {
-    const token = replacements[underlyingsIronBank.output[idx].output] || underlyingsIronBank.output[idx].output;
-    const amount = BigNumber(cashVal.output).plus(totalBorrowsIB.output[idx].output).minus(totalReservesIB.output[idx].output)
-    sdk.util.sumSingleBalance(balances, token, amount.toFixed(0));
-  });
 
   // --- Grab the accumulated on CRETH2 (ETH balance and update proper balances key) ---
   const accumCRETH2 = (
@@ -211,80 +174,6 @@ const bscBorrowed = async (timestamp, ethBlock, chainBlocks) => {
   return lending(block, "bsc", true)
 }
 
-const fantomToken = "0x4e15361fd6b4bb609fa63c81a2be19d873717870";
-const fantomTvl = async (timestamp, ethBlock, chainBlocks) => {
-  const block = chainBlocks["fantom"]; // req for the block type
-  let balances = {};
-
-  let tokens_fantom = (
-    await utils.fetchURL(
-      "https://api.cream.finance/api/v1/crtoken?comptroller=fantom"
-    )
-  ).data;
-
-  let cashValues = (
-    await sdk.api.abi.multiCall({
-      block,
-      calls: tokens_fantom.map((token) => ({ target: token.token_address })),
-      abi: abiCerc20["getCash"],
-      chain: "fantom",
-    })
-  ).output;
-
-  let underlyings = (
-    await sdk.api.abi.multiCall({
-      block,
-      calls: tokens_fantom.map((token) => ({ target: token.token_address })),
-      abi: abiCerc20["underlying"],
-      chain: "fantom",
-    })
-  ).output;
-
-  const fantomAddr = await transformFantomAddress()
-  cashValues.map((cashVal, idx) => {
-    const tokenAddr = underlyings[idx].output === "0x924828a9Fb17d47D0eb64b57271D10706699Ff11" ? "0xb753428af26e81097e7fd17f40c88aaa3e04902c" : fantomAddr(underlyings[idx].output); //SFI
-    sdk.util.sumSingleBalance(balances, tokenAddr, cashVal.output);
-  });
-
-  const stakerCreamAddressInFantom =
-    "0x0abad588c5490eee5850693e16bb6de9d60bdb6c";
-  const fantom_staking_service_endpoint = "https://xapi3.fantom.network/api";
-  const graphQLClient = new GraphQLClient(fantom_staking_service_endpoint);
-
-  const query = gql`
-    query StakerByAddress($address: Address!) {
-      staker(address: $address) {
-        id
-        stakerAddress
-        totalStake
-        stake
-        delegatedMe
-        createdEpoch
-        createdTime
-        downtime
-        lockedUntil
-        isActive
-        isOffline
-        stakerInfo {
-          name
-          website
-          contact
-          logoUrl
-          __typename
-        }
-        __typename
-      }
-    }
-  `;
-
-  const fantomStakingData = await graphQLClient.request(query, {
-    address: stakerCreamAddressInFantom,
-  });
-  sdk.util.sumSingleBalance(balances, fantomToken, fantomStakingData.staker.totalStake)
-
-  return balances;
-};
-
 module.exports = {
   timetravel: false, // bsc and fantom api's for staked coins can't be queried at historical points
   start: 1599552000, // 09/08/2020 @ 8:00am (UTC)
@@ -296,10 +185,5 @@ module.exports = {
     borrowed: bscBorrowed
     //getCompoundV2Tvl("0x589de0f0ccf905477646599bb3e5c622c84cc0ba", "bsc", addr=>`bsc:${addr}`,  "0x1Ffe17B99b439bE0aFC831239dDECda2A790fF3A", "0xbb4cdb9cbd36b01bd1cbaebf2de08d9173bc095c", true),
   },
-  fantom: {
-    tvl: fantomTvl,
-    borrowed: getCompoundV2Tvl("0x4250a6d3bd57455d7c6821eecb6206f507576cd2", "fantom", addr=>`fantom:${addr}`, undefined, undefined, true)
-  },
   polygon:compoundExports("0x20ca53e2395fa571798623f1cfbd11fe2c114c24", "polygon"),
-  avalanche:compoundExports("0x2eE80614Ccbc5e28654324a66A396458Fa5cD7Cc", "avax")
 };
